task router and controller handles jwt and user

+++ controllers/taskController.js
@@ -1,10 +1,13 @@
 const Task = require('../models/taskModel')
+const User = require('../models/userModel')
 const asyncHandler = require('express-async-handler')
 
-const getTasks = asyncHandler( async (req,res) =>{
-    const tasks = await Task.find({})
+//so what do we do here? token? id
 
-    res.status(200).send(tasks)
+const getTasks = asyncHandler( async (req,res) =>{
+    const tasks = await Task.find({user: req.user.id })
+
+    res.status(200).json(tasks)
 })
 
 const addTask = asyncHandler( async (req,res) =>{
@@ -15,7 +18,8 @@ const addTask = asyncHandler( async (req,res) =>{
     }
     const task = await Task.create({
         text,
-        complete
+        complete,
+        user: req.user.id,
     })
     if(task){
         res.status(200).send(task)
@@ -23,29 +27,68 @@ const addTask = asyncHandler( async (req,res) =>{
 })
 
 const updateTask = asyncHandler( async (req,res) =>{
-    const {text, complete} = req.body;
-    const updatedTask = await Task.findByIdAndUpdate({_id: req.params.id},{
-        text,
-        complete
-    },{new:true}) 
+    // const {text, complete} = req.body;
+    const task = await Task.findById(req.params.id)
 
-    if(updatedTask){
-        res.status(200).send(updatedTask)
+    if(!task){
+        res.status(400)
+        throw new Error('task not found')
     }
+
+    if(!req.user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    if(task.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('user not authorized')
+    }
+
+    //how i wrote it
+    // const updatedTask = await Task.findByIdAndUpdate({_id: req.params.id},{
+    //     text,
+    //     complete
+    // },{new:true})
+    const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body,{
+        new: true,
+    }) 
+
+
+    res.status(200).json(updatedTask)
     
-    res.status(400).send(`task not found`)
-    throw new Error('no task with that id')
 })
 
 const deleteTask = asyncHandler( async (req,res) =>{
-    const deletedTask = await Task.findByIdAndDelete(req.params.id) 
+    const task = await Task.findById(req.params.id)
 
-    if(deletedTask){
-        res.status(200).send(`deleted task ${id}`)
+    if(!task){
+        res.status(400)
+        throw new Error('task not found')
     }
+
+    if(!req.user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    if(task.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('user not authorized')
+    }
+
+    // await task.remove();
+     await Task.findByIdAndDelete(req.params.id)
+
+    res.status(200).json({ id: req.params.id })
+    // const deletedTask = await Task.findByIdAndDelete(req.params.id) 
+
+    // if(deletedTask){
+    //     res.status(200).send(`deleted task ${id}`)
+    // }
     
-    res.status(400).send(`task not found`)
-    throw new Error('no task with that id')
+    // res.status(400).send(`task not found`)
+    // throw new Error('no task with that id')
 })
